% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsicKappa.R
\name{computeStat}
\alias{computeStat}
\title{Compute Statistics}
\usage{
computeStat(n1, n2, alpha)
}
\arguments{
\item{n1}{integer}

\item{n2}{integer}

\item{alpha}{one-sided significance level}
}
\description{
Compute Statistics
}
