% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDescription.R
\docType{package}
\name{intrinsicKappa-package}
\alias{intrinsicKappa-package}
\title{Sample Size Planning Based on Intrinsic Kappa Value}
\description{
Providing functions for the computation of the intrinsic kappa value.
}
\references{
R. Sanchez-Marquez, F. Gerhorst and D. Schindler (2023) "Effectiveness of quality
inspections of attributive characteristics – A novel and practical method for estimating
the “intrinsic” value of kappa based on alpha and beta statistics." \emph{Computers & Industrial Engineering}, 109006.
}
\seealso{
For the computation of the intrinsic kappa value, see \code{\link{intrinsicKappa}}.
}
\author{
David Schindler \email{dv.schindler@gmail.com}, Rafael Sanchez-Marquez, Frank Gerhorst
}
